Move mana algorithms to haiku-common
